indev

+++ src/main/kotlin/xyz/teodorowicz/pm/repository/UserRepository.kt
@@ -1,11 +1,13 @@
 package xyz.teodorowicz.pm.repository
 
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import xyz.teodorowicz.pm.entity.User
 
 @Repository
-interface UserRepository : JpaRepository<User, Int> {
+interface UserRepository : JpaRepository<User, Long> {
     fun findByEmail(email: String): User?
     fun existsByEmail(email: String): Boolean
 }
